rpc: move grpc_test from test/ to grpc/
